refactor(domain): split info.rs into module directory
